Inpaintings.jl Documentation

Inpaintings.jl Documentation

Inpaintings.jl provides functions that perform a similar task to MATLAB's inpaint_nans function. (MATLAB's inpaint_nans was originally written by John d'Errico and is available on the MathWorks File Exchange website.)

inpaint_nans takes a vector or a matrix as input and fills ("inpaints") the NaNs by solving a simple 1D or 2D PDE.

So far Inpaintings.jl only replicates John d'Errico's method 0 well, which is well suited for filling NaNs in a "diffusive" way. Methods 1 and 3 are on their way. Other methods will come next.

Usage

Use the package as you would any Julia package

The command inpaint_nans(A) will fill the NaNs of an Array A (that has some NaNs):

# Making a test matrix A with some NaNs by replacing values from fullA
n = 10
fullA = float(collect(1:n) * collect(1:n)')
A = copy(fullA)
A[1:5, 1:5] .= NaN # replace some values with some NaNs
B = inpaint(A, NaN)
B ≈ fullA

# output

true

Functions

Inpaintings.inpaint — Method.

inpaint(f, A)

Inpaints values of A for which f(A) == true.

Examples

julia> A = float((1:5)*(1:5)') ; A[1:2, 1:2] .= 999 ; A
5×5 Array{Float64,2}:
 999.0  999.0   3.0   4.0   5.0
 999.0  999.0   6.0   8.0  10.0
   3.0    6.0   9.0  12.0  15.0
   4.0    8.0  12.0  16.0  20.0
   5.0   10.0  15.0  20.0  25.0

julia> inpaint(x -> x == 999, A)
5×5 Array{Float64,2}:
 1.0   2.0   3.0   4.0   5.0
 2.0   4.0   6.0   8.0  10.0
 3.0   6.0   9.0  12.0  15.0
 4.0   8.0  12.0  16.0  20.0
 5.0  10.0  15.0  20.0  25.0

julia> A = float((1:5)*(1:5)') ; A[1:2, [1, end]] .= NaN ; A
5×5 Array{Float64,2}:
 NaN     2.0   3.0   4.0  NaN
 NaN     4.0   6.0   8.0  NaN
   3.0   6.0   9.0  12.0   15.0
   4.0   8.0  12.0  16.0   20.0
   5.0  10.0  15.0  20.0   25.0

julia> inpaint(isnan, A)
5×5 Array{Float64,2}:
 1.0   2.0   3.0   4.0   5.0
 2.0   4.0   6.0   8.0  10.0
 3.0   6.0   9.0  12.0  15.0
 4.0   8.0  12.0  16.0  20.0
 5.0  10.0  15.0  20.0  25.0

inpaint(f, A; method=1, cycledims=[])

With optional arguments, you can chose the inpainting method and if dimensions are cyclic.

julia> A = float((1:5)*(1:5)') ; A[1:2, [1, end]] .= NaN ; A
5×5 Array{Float64,2}:
 NaN     2.0   3.0   4.0  NaN
 NaN     4.0   6.0   8.0  NaN
   3.0   6.0   9.0  12.0   15.0
   4.0   8.0  12.0  16.0   20.0
   5.0  10.0  15.0  20.0   25.0

julia> inpaint(A, cycledims=[2])
5×5 Array{Float64,2}:
 2.13971   2.0   3.0   4.0   3.86029
 3.50735   4.0   6.0   8.0   8.49265
 3.0       6.0   9.0  12.0  15.0
 4.0       8.0  12.0  16.0  20.0
 5.0      10.0  15.0  20.0  25.0

source

Inpaintings.inpaint — Method.

inpaint(A, missing)

Inpaints missing values. Should be the same as inpaint(A).

source

Inpaintings.inpaint — Method.

inpaint(A, value_to_fill)

Inpaints value_to_fill values (can be NaN). Useful in the case data was generated by filling missing values with an arbitrary chosen one.

Examples

julia> A = float((1:5)*(1:5)') ; A[1:2, 1:2] .= 999 ; A
5×5 Array{Float64,2}:
 999.0  999.0   3.0   4.0   5.0
 999.0  999.0   6.0   8.0  10.0
   3.0    6.0   9.0  12.0  15.0
   4.0    8.0  12.0  16.0  20.0
   5.0   10.0  15.0  20.0  25.0

julia> inpaint(A, 999)
5×5 Array{Float64,2}:
 1.0   2.0   3.0   4.0   5.0
 2.0   4.0   6.0   8.0  10.0
 3.0   6.0   9.0  12.0  15.0
 4.0   8.0  12.0  16.0  20.0
 5.0  10.0  15.0  20.0  25.0

julia> A = float((1:5)*(1:5)') ; A[1:2, 1:2] .= NaN ; A
5×5 Array{Float64,2}:
 NaN    NaN     3.0   4.0   5.0
 NaN    NaN     6.0   8.0  10.0
   3.0    6.0   9.0  12.0  15.0
   4.0    8.0  12.0  16.0  20.0
   5.0   10.0  15.0  20.0  25.0

julia> inpaint(A, NaN)
5×5 Array{Float64,2}:
 1.0   2.0   3.0   4.0   5.0
 2.0   4.0   6.0   8.0  10.0
 3.0   6.0   9.0  12.0  15.0
 4.0   8.0  12.0  16.0  20.0
 5.0  10.0  15.0  20.0  25.0

source

Inpaintings.inpaint — Method.

inpaint(A)

Inpaints NaN values if the elements of A are all float (i.e., eltype(A) <: AbstractFloat).

Example

julia> A = float((1:5)*(1:5)') ; A[1:2, 1:2] .= NaN ; A
5×5 Array{Float64,2}:
 NaN    NaN     3.0   4.0   5.0
 NaN    NaN     6.0   8.0  10.0
   3.0    6.0   9.0  12.0  15.0
   4.0    8.0  12.0  16.0  20.0
   5.0   10.0  15.0  20.0  25.0

julia> inpaint(A)
5×5 Array{Float64,2}:
 1.0   2.0   3.0   4.0   5.0
 2.0   4.0   6.0   8.0  10.0
 3.0   6.0   9.0  12.0  15.0
 4.0   8.0  12.0  16.0  20.0
 5.0  10.0  15.0  20.0  25.0

source

Inpaintings.inpaint — Method.

inpaint(A)

Inpaints missing values if A contains some.

Inspired by MATLAB's inpaint_nans's (by John d'Errico: link).

Example

julia> A = convert(Array{Union{Float64, Missing},2}, (1:5)*(1:5)') ; A[1:2, 1:2] .= missing ; A
5×5 Array{Union{Missing, Float64},2}:
  missing    missing   3.0   4.0   5.0
  missing    missing   6.0   8.0  10.0
 3.0        6.0        9.0  12.0  15.0
 4.0        8.0       12.0  16.0  20.0
 5.0       10.0       15.0  20.0  25.0

julia> inpaint(A)
5×5 Array{Union{Missing, Float64},2}:
 1.0   2.0   3.0   4.0   5.0
 2.0   4.0   6.0   8.0  10.0
 3.0   6.0   9.0  12.0  15.0
 4.0   8.0  12.0  16.0  20.0
 5.0  10.0  15.0  20.0  25.0

source

Inpaintings.inpaint_method0 — Method.

inpaint_method0(f, A::Vector)

Inpaints values in A that f gives true on by solving a simple diffusion PDE. The partial differential equation (PDE) is defined by the standard Laplacian, Δ = ∇^2. Inspired by MATLAB's inpaint_nans's method 0 for vectors (by John d'Errico). See https://www.mathworks.com/matlabcentral/fileexchange/4551-inpaint_nans.

source

Inpaintings.inpaint_method1 — Method.

inpaint_method1(f, A::Array, cycledims=Int64[])

Inpaints values in A that f gives true on by solving a simple diffusion PDE. Default method for inpaint. The partial differential equation (PDE) is defined by the standard Laplacian, Δ = ∇^2. Inspired by MATLAB's inpaint_nans's method 0 for matrices (by John d'Errico). See https://www.mathworks.com/matlabcentral/fileexchange/4551-inpaint_nans. The discrete stencil used for Δ looks like

      ┌───┐
      │ 1 │
      └─┬─┘
        │
┌───┐ ┌─┴─┐ ┌───┐
│ 1 ├─┤-4 ├─┤ 1 │
└───┘ └─┬─┘ └───┘
        │
      ┌─┴─┐
      │ 1 │
      └───┘

By default, the stencil is not applied at the borders. Instead, its 1D component,

┌───┐ ┌───┐ ┌───┐
│ 1 ├─┤-2 ├─┤ 1 │
└───┘ └───┘ └───┘

is applied where it fits at the borders. However, the user can supply a list of dimensions that should be considered cyclic. In this case, the sentil will be used also at the borders and "jump" to the other side. This is particularly useful for, e.g., world maps with longitudes spanning the entire globe.

source

Inpaintings.list_neighbors — Method.

list_neighbors(A, idx, neighbors)

Lists all the neighbors of the indices in idx in Array A. Neighbors already in idx are accepted. The argument neighnors contains a list of the neighbors about the origin coordinate (0, 0, ...). In other words, it is a Vector of CartesianIndex such that the direct neighbors of index i are given by i + n for n in neighbors. Inspired by MATLAB's inpaint_nans's identify_neighbors (by John d'Errico). See https://www.mathworks.com/matlabcentral/fileexchange/4551-inpaint_nans.

source