Inpaintings.jl Documentation

Inpaintings.jl Documentation

Inpaintings.jl provides functions that perform a similar task to MATLAB's inpaint_nans function. (MATLAB's inpaint_nans was originally written by John d'Errico and is available on the MathWorks File Exchange website.)

inpaint_nans takes a vector or a matrix as input and fills ("inpaints") the NaNs by solving a simple 1D or 2D PDE.

So far Inpaintings.jl only replicates John d'Errico's method 0 well, which is well suited for filling NaNs in a "diffusive" way. Methods 1 and 3 are on their way. Other methods will come next.

Usage

Use the package as you would any Julia package

The command inpaint_nans(A) will fill the NaNs of an Array A (that has some NaNs):

# Making a test matrix A with some NaNs by replacing values from fullA
n = 10
fullA = float(collect(1:n) * collect(1:n)')
A = copy(fullA)
A[1:5, 1:5] .= NaN # replace some values with some NaNs
B = inpaint_nans(A)
B ≈ fullA

# output

true

Functions

Inpaintings.inpaint_nans — Function.

inpaint_nans(A, method=0)

Inpaints NaN values by solving a diffusion PDE for the standard Laplacian. Inspired by MATLAB's inpaint_nans's (by John d'Errico). See https://www.mathworks.com/matlabcentral/fileexchange/4551-inpaint_nans. Currently only method 0 is implemented.

source

Inpaintings.inpaint_nans_method0 — Method.

inpaint_nans_method0(A::Vector)

Inpaints NaN values by solving a diffusion PDE for the standard Laplacian. Inspired by MATLAB's inpaint_nans's method 0 for vectors (by John d'Errico). See https://www.mathworks.com/matlabcentral/fileexchange/4551-inpaint_nans.

source

Inpaintings.inpaint_nans_method0 — Method.

inpaint_nans_method0(A::Array{T,2}) where T

Inpaints NaN values by solving a diffusion PDE for the standard Laplacian. Inspired by MATLAB's inpaint_nans's method 0 for matrices (by John d'Errico). See https://www.mathworks.com/matlabcentral/fileexchange/4551-inpaint_nans. The discrete stencil used for ∇² looks like

            ┌───┐
            │ 1 │
            └─┬─┘
              │
      ┌───┐ ┌─┴─┐ ┌───┐
      │ 1 ├─┤-4 ├─┤ 1 │
      └───┘ └─┬─┘ └───┘
              │
            ┌─┴─┐
            │ 1 │
            └───┘

The stencil is not applied at the coreners, but its 1D components,

   ┌───┐
   │ 1 │
   └─┬─┘
     │
   ┌─┴─┐        ┌───┐ ┌───┐ ┌───┐
   │-2 │   or   │ 1 ├─┤-2 ├─┤ 1 │
   └─┬─┘        └───┘ └───┘ └───┘
     │
   ┌─┴─┐
   │ 1 │
   └───┘

are applied at borders.

source

Inpaintings.inpaint_nans_method1 — Method.

inpaint_nans_method1(A::Array{T,2}) where T

Inpaints NaN values by solving a diffusion PDE for the standard Laplacian. Inspired by MATLAB's inpaint_nans's method 0 for matrices (by John d'Errico). See https://www.mathworks.com/matlabcentral/fileexchange/4551-inpaint_nans. The discrete stencil used for ∇² looks like

            ┌───┐
            │ 1 │
            └─┬─┘
              │
      ┌───┐ ┌─┴─┐ ┌───┐
      │ 1 ├─┤-4 ├─┤ 1 │
      └───┘ └─┬─┘ └───┘
              │
            ┌─┴─┐
            │ 1 │
            └───┘

The stencil is not applied at the coreners, but its 1D components,

   ┌───┐
   │ 1 │
   └─┬─┘
     │
   ┌─┴─┐        ┌───┐ ┌───┐ ┌───┐
   │-2 │   or   │ 1 ├─┤-2 ├─┤ 1 │
   └─┬─┘        └───┘ └───┘ └───┘
     │
   ┌─┴─┐
   │ 1 │
   └───┘

are applied at borders.

source

Inpaintings.inpaint_nans_method3 — Method.

inpaint_nans_method3(A::Array{T,2}) where T

Inpaints NaN values by solving a diffusion PDE for ∇⁴: Inspired by MATLAB's inpaint_nans's method 3 for matrices (by John d'Errico). The discrete stencil used for ∇⁴ looks like

            ┌───┐
            │ 1 │
            └─┬─┘
              │
      ┌───┐ ┌─┴─┐ ┌───┐
      │ 2 ├─┤-8 ├─┤ 2 │
      └─┬─┘ └─┬─┘ └─┬─┘
        │     │     │
┌───┐ ┌─┴─┐ ┌─┴─┐ ┌─┴─┐ ┌───┐
│ 1 ├─┤-8 ├─┤20 ├─┤-8 ├─┤ 1 │
└───┘ └─┬─┘ └─┬─┘ └─┬─┘ └───┘
        │     │     │
      ┌─┴─┐ ┌─┴─┐ ┌─┴─┐
      │ 2 ├─┤-8 ├─┤ 2 │
      └───┘ └─┬─┘ └───┘
              │
            ┌─┴─┐
            │ 1 │
            └───┘

The stencil is actually constructed from its 1st order 1D components,

   ┌───┐
   │ 1 │
   └─┬─┘
     │
   ┌─┴─┐        ┌───┐ ┌───┐ ┌───┐
   │-2 │   or   │ 1 ├─┤-2 ├─┤ 1 │
   └─┬─┘        └───┘ └───┘ └───┘
     │
   ┌─┴─┐
   │ 1 │
   └───┘

which are applied at the borders.

source

Inpaintings.inpaint_nans_method6 — Method.

inpaint_nans_method6(A::Array{T,2}) where T

Inpaints NaN values by solving a diffusion PDE for a diagonally filled Laplacian:

\[\mathbf{D}_{xy}^{2} = \begin{bmatrix}0.25&0.5&0.25\\0.5&-3&0.5\\0.25&0.5&0.25\end{bmatrix}\]

Inspired by MATLAB's inpaint_nans's method 0 for matrices (by John d'Errico). See https://www.mathworks.com/matlabcentral/fileexchange/4551-inpaint_nans.

source

Inpaintings.list_neighbors — Method.

list_neighbors(A, idx, neighbors)

Lists all the neighbors of the indices in idx in Array A. Neighbors already in idx are accepted. The argument neighnors contains a list of the neighbors about the origin coordinate (0, 0, ...). In other words, it is a Vector of CartesianIndex such that the direct neighbors of index i are given by i + n for n in neighbors. Inspired by MATLAB's inpaint_nans's identify_neighbors (by John d'Errico). See https://www.mathworks.com/matlabcentral/fileexchange/4551-inpaint_nans.

source