Inpaintings.jl Documentation

Inpaintings.jl Documentation

Inpaintings.jl provides functions that perform a similar task to MATLAB's inpaint_nans function. (MATLAB's inpaint_nans was originally written by John d'Errico and is available on the MathWorks File Exchange website.)

inpaint_nans takes a vector or a matrix as input and fills ("inpaints") the NaNs by solving a simple 1D or 2D PDE.

So far Inpaintings.jl only replicates John d'Errico's method 0 well, which is well suited for filling NaNs in a "diffusive" way. Methods 1 and 3 are on their way. Other methods will come next.

Usage

Use the package as you would any Julia package

The command inpaint_nans(A) will fill the NaNs of an Array A (that has some NaNs):

# Making a test matrix A with some NaNs by replacing values from fullA
n = 10
fullA = float(collect(1:n) * collect(1:n)')
A = copy(fullA)
A[1:5, 1:5] .= NaN # replace some values with some NaNs
B = inpaint_nans(A)
B ≈ fullA

# output

true

Functions

Inpaintings.inpaint_nans — Function.

inpaint_nans(A, method=0)

Inpaints NaN values by solving a diffusion PDE for the standard Laplacian. Inspired by MATLAB's inpaint_nans's (by John d'Errico). See https://www.mathworks.com/matlabcentral/fileexchange/4551-inpaint_nans. Currently only method 0 is implemented.

source